Removed Node (express, socket) stuff. Revamped for python pwnage.
I removed all node related things. Not going to work on the frontend just yet.
Also, user auth has be moved to use mongodb and Flask-Login rather than basic http auth. Going to search for a way to enable basic/digest auth for API request once that is built.
Lastly, the database models for the music library has been started. These models are not definite, and may not even be used.